requirements.txt
# J-Quants株価予測システム 必要パッケージ

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# API通信
requests>=2.28.0

# 機械学習
scikit-learn>=1.3.0
xgboost>=2.0.0

# データ可視化
matplotlib>=3.6.0
seaborn>=0.12.0

# 設定管理
python-dotenv>=1.0.0
pyyaml>=6.0.0

# 開発・テスト用
jupyter>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
black>=23.0.0
flake8>=6.0.0
coverage>=7.0.0
